fix: move _CURRENT_TASK ContextVar to separate module to avoid dill serialization issue
dill serializes module dicts when pickling classes. Having a ContextVar
in flux/task.py made the task class unpicklable. Moving it to
flux/_task_context.py and using a local import in task.__call__
keeps it out of the module dict that dill traverses.
