+++ examples/ai/sample_data/flux_overview.md
@@ -1,0 +1,12 @@
+# Overview
+
+Flux is a distributed workflow orchestration engine designed to build stateful and fault-tolerant workflows in Python. It provides a robust framework for creating, managing, and executing complex workflows with built-in support for error handling, retries, and state management.
+
+## What is Flux?
+
+Flux is a Python-based workflow orchestration system that allows you to:
+- Define workflows as Python code using decorators
+- Break down complex processes into manageable tasks
+- Handle failures and retries automatically
+- Maintain state across workflow executions
+- Scale from local development to distributed systems
