+++ flux/_task_context.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from contextvars import ContextVar
+
+_CURRENT_TASK: ContextVar[tuple[str, str] | None] = ContextVar("current_task", default=None)
